+++ Actions/DetermineArtifactUrl/DetermineArtifactUrl.ps1
@@ -8,10 +8,34 @@
 DownloadAndImportBcContainerHelper
 #endregion
 
-#region Action: Determine artifacts to use
+#region Action: Determine artifacts to use based on buildAuthContext or settings
+$artifactUrl = $null
 $settings = $env:Settings | ConvertFrom-Json | ConvertTo-HashTable
-$settings = AnalyzeRepo -settings $settings -project $project -doNotCheckArtifactSetting -doNotIssueWarnings
-$artifactUrl = DetermineArtifactUrl -projectSettings $settings
+if ($env:Secrets) {
+    $secrets = $env:Secrets | ConvertFrom-Json | ConvertTo-HashTable
+    if ($secrets.ContainsKey('buildAuthContext') -and $settings.buildEnvironmentName) {
+        $buildAuthContext = [System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String($secrets.buildAuthContext))
+        $authContextHT = $buildAuthContext | ConvertFrom-Json | ConvertTo-HashTable
+        $environmentName = $settings.buildEnvironmentName
+        if ($environmentName -notlike 'https://*') {
+            $authContext = New-BcAuthContext @authContextHT
+            $bcEnvironment = Get-BcEnvironments -bcAuthContext $authContext | Where-Object { $_.name -eq $environmentName -and $_.type -eq "Sandbox" }
+            if (-not $bcEnvironment) {
+                throw "Business Central online environment '$environmentName' not found"
+            }
+            $settings.useCompilerFolder = $true
+            $bcBaseApp = Get-BcPublishedApps -bcAuthContext $authContext -environment $environmentName | Where-Object { $_.Name -eq "Base Application" -and $_.state -eq "installed" }
+            if (-not $bcBaseApp) {
+                throw "Base Application not found in environment '$environmentName'"
+            }
+            $artifactUrl = Get-BCArtifactUrl -type Sandbox -country $bcEnvironment.countryCode -version $bcBaseApp.Version -select Closest
+        }
+    }
+}
+if (!$artifactUrl) {
+    $settings = AnalyzeRepo -settings $settings -project $project -doNotCheckArtifactSetting -doNotIssueWarnings
+    $artifactUrl = DetermineArtifactUrl -projectSettings $settings
+}
 $artifactCacheKey = ''
 if ($settings.useCompilerFolder) {
     $artifactCacheKey = $artifactUrl.Split('?')[0]