+++ Actions/RunPipeline/RunPipeline.ps1
@@ -86,7 +86,7 @@ try {
 
     $appBuild = $settings.appBuild
     $appRevision = $settings.appRevision
-    'licenseFileUrl','codeSignCertificateUrl','codeSignCertificatePassword','keyVaultCertificateUrl','*keyVaultCertificatePassword','keyVaultClientId','gitHubPackagesContext','applicationInsightsConnectionString' | ForEach-Object {
+    'licenseFileUrl','codeSignCertificateUrl','codeSignCertificatePassword','keyVaultCertificateUrl','*keyVaultCertificatePassword','keyVaultClientId','gitHubPackagesContext','applicationInsightsConnectionString','buildAuthContext' | ForEach-Object {
         # Secrets might not be read during Pull Request runs
         if ($secrets.Keys -contains $_) {
             $value = [System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String($secrets."$_"))
@@ -100,6 +100,18 @@ try {
     }
 
     $analyzeRepoParams = @{}
+
+    $authContext = $null
+    $environmentName = ""
+    if ($buildAuthContext -and $settings.buildEnvironmentName) {
+        $authContextHT = $buildAuthContext | ConvertFrom-Json | ConvertTo-HashTable
+        $environmentName = $settings.buildEnvironmentName
+        if ($environmentName -notlike 'https://*') {
+            $authContext = New-BcAuthContext @authContextHT
+            # When using online environment for builds, use CompilerFolder to avoid creating a container
+            $settings.useCompilerFolder = $true
+        }
+    }
 
     if ($artifact) {
         # Avoid checking the artifact setting in AnalyzeRepo if we have an artifactUrl
@@ -279,9 +291,6 @@ try {
             Write-Host "::endgroup::"
         }
     }
-    $authContext = $null
-    $environmentName = ""
-    $CreateRuntimePackages = $false
 
     if ($settings.versioningStrategy -eq -1) {
         $artifactVersion = [Version]$settings.artifact.Split('/')[4]
@@ -522,7 +531,6 @@ try {
         -buildArtifactFolder $buildArtifactFolder `
         -pageScriptingTestResultsFile (Join-Path $buildArtifactFolder 'PageScriptingTestResults.xml') `
         -pageScriptingTestResultsFolder (Join-Path $buildArtifactFolder 'PageScriptingTestResultDetails') `
-        -CreateRuntimePackages:$CreateRuntimePackages `
         -appBuild $appBuild -appRevision $appRevision `
         -uninstallRemovedApps