+++ RELEASENOTES.md
@@ -9,6 +9,30 @@
 - Issue 1657 When no files modified on Git, deployment fails
 - Issue 1530 Dependency Field Service Integration does not get published in container while Installing apps
 
+### New Workflow specific settings
+
+- `buildEnvironmentName` - should be the Business Central environment name (from the Admin Center) for the environment you want to use if you want to run tests using online environments. Note that you also need to create a [buildAuthContext](https://aka.ms/algosecrets#buildAuthContext) secret, which gives access to this environment.
+
+### Use online environment during build for running tests
+
+By creating a [build authentication context](https://aka.ms/algosecrets#buildAuthContext) secret and setting the [buildEnvironmentName](https://aka.ms/algosettings#buildEnvironmentName) for a specific branch and a specific workflow, will cause AL-Go for GitHub to use this environment for running tests. An example of how to define these settings could be:
+
+```json
+  "useCompilerFolder": true,
+  "doNotPublishApps": true,
+  "conditionalSettings": [
+    {
+      "workflows": [ "CICD" ],
+      "branches": [ "main" ],
+      "settings": {
+        "doNotPublishApps": false,
+        "buildEnvironmentName": "cicd",
+        "buildAuthContextSecretName": "cicdAuthContext"
+      }
+    }
+  ]
+```
+
 ## v7.0
 
 ### Issues
@@ -881,7 +905,7 @@ Setting the repo setting "runs-on" to "Ubuntu-latest", followed by running Updat
 - New setting: **DoNotPublishApps** - setting this to true causes the workflow to skip publishing, upgrading and testing the app to improve performance.
 - New setting: **ConditionalSettings** to allow to use different settings for specific branches. Example:
 
-```
+```json
     "ConditionalSettings": [
         {
             "branches": [